+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Upgrade to Alchemy v1.4.0
 * Add edit template for the `Object` field type
 * Add settings editor element
+* Make `readonly` attribute work on `al-field` and `al-form` elements
 
 ## 0.2.10 (2024-01-15)